+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 # PayPal iOS SDK Release Notes
 
 ## unreleased
+* PayPalNativePayments
+  * Fix bug where newly selected shipping method was not reflected in the Paysheet UI
 * Card
   * Rename `Card.init()` parameter from `cardHolderName` to `cardholderName`
   * Remove unnecessary `Card.expiry` property